delete, modify still frontend not working but in little stable form

+++ client/components/PersonHolder.js
@@ -6,42 +6,48 @@ import personService from '../services/persons'
 
 const PersonHolder = ({
   persons, setPerson, setPersons, setErrorMessage, setNotificationType,
-}) => (
+}) => {
+  if (!persons) return null
 
-  persons.map((person) => (
-    <p
-      key={person.id}
-    >
-      {person.name}
-      {' '}
-      {person.number}
-      {' '}
-      <button
+  console.log(persons)
+
+  return (
+
+    persons.map((person) => (
+      <p
         key={person.id}
-        onClick={() => {
-          if (window.confirm(`Delete ${person.name} ?`)) {
-            personService
-              .deletePerson(person.id)
-              .catch((error) => {
-                console.log('error', error)
-
-                setErrorMessage(`Information of ${person.name} has already been removed from server`)
-
-                console.log(typeof setErrorMessage)
-                setNotificationType('error')
-              })
-
-            const t = persons.filter((p) => p.id !== person.id)
-            setPersons(t)
-            const temp = <PersonHolder persons={t} setPerson={setPerson} setPersons={setPersons} />
-            setPerson(temp)
-          }
-        }}
       >
-        delete
-      </button>
-    </p>
-  ))
-)
+        {person.name}
+        {' '}
+        {person.number}
+        {' '}
+        <button
+          key={person.id}
+          onClick={() => {
+            if (window.confirm(`Delete ${person.name} ?`)) {
+              personService
+                .deletePerson(person.id)
+                .catch((error) => {
+                  console.log('error', error)
+
+                  setErrorMessage(`Information of ${person.name} has already been removed from server`)
+
+                  console.log(typeof setErrorMessage)
+                  setNotificationType('error')
+                })
+
+              const t = persons.filter((p) => p.id !== person.id)
+              setPersons(t)
+              const temp = <PersonHolder persons={t} setPerson={setPerson} setPersons={setPersons} />
+              setPerson(temp)
+            }
+          }}
+        >
+          delete
+        </button>
+      </p>
+    ))
+  )
+}
 
 export default PersonHolder
